refactor: isolate unix metadata handling
