+++ Cargo.toml
@@ -21,7 +21,6 @@ chrono = "0.4.41"
 glob = "0.3.1"
 clap = { version = "4.5.37", features = ["derive"] }
 colored_text = "0.5.1"
-nix = { version = "0.31.0", features = ["user"] }
 terminal_size = "0.4.4"
 config = "0.15.7"
 serde = { version = "1.0.219", features = ["derive"] }
@@ -29,6 +28,9 @@ dirs-next = "2.0.0"
 strip-ansi-escapes = "0.2.1"
 unicode-width = "0.2.0"
 ignore = "0.4.25"
+
+[target.'cfg(unix)'.dependencies]
+nix = { version = "0.31.0", features = ["user"] }
 
 [dev-dependencies]
 tempfile = "3.19.1"
