+++ src/platform/mod.rs
@@ -1,0 +1,7 @@
+//! Platform-specific filesystem metadata interpretation.
+
+#[cfg(unix)]
+mod unix;
+
+#[cfg(unix)]
+pub(crate) use unix::*;
